requirements.txt
fastapi
python-dotenv
uvicron
requests
os
